+++ Sources/KauppaTax/Model/TaxRate.swift
@@ -4,10 +4,10 @@ import KauppaCore
 /// supposed to be in percentages.
 public struct TaxRate: Mappable {
     /// Tax rate for any product.
-    public var general: Float = 0.0
+    public var general: Double = 0.0
     /// Tax rates for different categories of products in the country/region.
     /// (also called tax classes)
-    public var categories = [String: Float]()
+    public var categories = [String: Double]()
 
     public init() {}
